Fixed flash-of-unstyled-text (FOUT) (in an ugly way)
Not a very nice fix at all. The FOUT is due to the fact that the font is
loaded via javascript through typekit, and so until the font has been
loaded, a fallback font is shown. Typekit claims this should only happen
on Firefox, but for me it happens all the time.

The fix works by hiding (visibility: hidden) all typekit-typeset
elements until the font has been loaded. An extra piece of javascript
sets the html-class to wf-loading, before typekit does so, so that no
text is shown before the typekit js can be shown.